refactor: prepare for multiple Slack bots to operate independently in the same workspace
